# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "\/api\/distributeur\/depot": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un dépôt",
                "description": "Permet à un distributeur d'effectuer un dépôt sur le compte d'un client",
                "operationId": "b8ca574b94e040c7b7e99982a0a866d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/TransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dépôt effectué avec succès",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#\/components\/schemas\/Transaction"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dépôt effectué avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors du dépôt",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/distributeur\/retrait": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un retrait",
                "description": "Permet à un distributeur d'effectuer un retrait sur le compte d'un client",
                "operationId": "ed53faa5420f9f2fa75b8edfd6c60070",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/TransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Retrait effectué avec succès",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#\/components\/schemas\/Transaction"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Retrait effectué avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors du retrait",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/distributeur\/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Lister les transactions du distributeur",
                "description": "Récupère toutes les transactions effectuées par le distributeur connecté",
                "operationId": "05ec688faafa13ced1d9492351591378",
                "responses": {
                    "200": {
                        "description": "Transactions récupérées avec succès",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#\/components\/schemas\/Transaction"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transactions récupérées avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token invalide",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/distributeur\/transactions\/{transaction}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Afficher une transaction spécifique",
                "description": "Récupère les détails d'une transaction spécifique du distributeur",
                "operationId": "f442e5272f769e457767a3a59c4c6fdc",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction récupérée avec succès",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#\/components\/schemas\/Transaction"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/client\/solde": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer le solde du client",
                "description": "Récupère le solde actuel du compte du client connecté",
                "operationId": "fc06ddad4126dc4f93e8cceed5690b08",
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#\/components\/schemas\/SoldeResponse"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde récupéré avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun compte trouvé",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/client\/transfert": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un transfert",
                "description": "Permet à un client d'effectuer un transfert vers un autre numéro",
                "operationId": "d30c3b396cc10cf30bf127cc6b927223",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/TransferRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transfert effectué avec succès",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#\/components\/schemas\/TransferResponse"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfert effectué avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun compte trouvé",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors du transfert",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/client\/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer les transactions du client",
                "description": "Récupère l'historique des transactions du client connecté",
                "operationId": "ef241b9fc6a36f0b5314ea19b99d8ff4",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions récupérées avec succès",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#\/components\/schemas\/Transaction"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transactions récupérées avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun compte trouvé",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "info": {
        "title": "OM Pay API",
        "version": "1.0.0",
        "description": "API de paiement mobile pour OM Pay - Système de paiement et transfert d'argent",
        "contact": {
            "email": "contact@ompay.com"
        }
    },
    "servers": [
        {
            "url": "https:\/\/om-pay-qdx2.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http:\/\/localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}